refactor(forms): move control directive under api
Moves the control directive under api/ since it is part of the public
API. Also removes the interop abstract control from the public API
